==> 1_Friends_Analysis.ipynb <==
import os, json
import pandas as pd

path_to_json = '/home/vincenzo/research/eni/paper/twittergeneralinterface_bkp-2020.01.27/'
json_files = [pos_json for pos_json in os.listdir(path_to_json) if pos_json.startswith('friends.')]
#json_files = [pos_json for pos_json in os.listdir(path_to_json) if pos_json.startswith('friends.2020-01-28_15-25')]
print("Found",len(json_files),"files")

# here I define my pandas Dataframe with the columns I want to get from the json
friends_df = pd.DataFrame(columns=['user_screen_name', 'screen_name', 'location', 'friends_count', 'listed_count', 'followers_count', 'created_at', 'lang', 'statuses_count', 'description'])

# we need both the json and an index number so use enumerate()
cnt=0
for index, js in enumerate(json_files):
    with open(os.path.join(path_to_json, js)) as json_file:
        print(js, end=" ")
        cnt_line=0
        for line in json_file:
            json_text = json.loads(line)

            # here you need to know the layout of your json and each json has to have
            # the same structure (obviously not the structure I have here)
            user_screen_name = json_text['user_screen_name']
            screen_name = json_text['screen_name']
            location = json_text['location']
            friends_count = json_text['friends_count']
            listed_count = json_text['listed_count']
            followers_count = json_text['followers_count']
            created_at = json_text['created_at']
            lang = json_text['lang']
            statuses_count = json_text['statuses_count']
            description = json_text['description']



            # here I push a list of data into a pandas DataFrame at row given by 'index'
            friends_df.loc[cnt] = [user_screen_name,
                                     screen_name,
                                     location,
                                     friends_count,
                                     listed_count,
                                     followers_count,
                                     created_at,
                                     lang,
                                     statuses_count,
                                     description]

            cnt+=1
            
            cnt_line+=1
        print("- Loaded",cnt_line,"users")

# now that we have the pertinent json data in our DataFrame let's look at it
#print(friends_df)

friends_df

unique_friends_df = friends_df.drop_duplicates()
unique_friends_df

# VALUTARE LA POSSIBILITA' DI USARE user_screen_name E screen_name COME INDICE DELLA TABELLA
# unique_friends_new_index_df = unique_friends_df.set_index(['user_screen_name', 'screen_name'])
# unique_friends_new_index_df

unique_friends_df.groupby("user_screen_name").agg('count')['friends_count']
# EXPECTED VALUES FROM OFFICIAL TWITTER ACCOUNTS:
# 
# ERGnow              3288
# EdisonNews          1602
# EnelGroupIT         5896                                                                                                        
# GruppoHera          606
# LandirenzoSpA       147
# TernaSpA            686
# cirgroup            424
# eni                 626
# falckrenewables     309
# gruppo_a2a          141
# gruppoiren          508
# innovatec_spa       282
# italgas             49
# snam                463


unique_friends_df.to_csv("friends_1st_level.csv")

unique_friends_df.to_json("friends_1st_level.json", orient='records')

# NUMERO COMPLESSIVO DI FRIENDS UNICI DI PRIMO LIVELLO
unique_friends_df['screen_name'].unique().shape
