==> 2_Friends_second_filter.ipynb <==
import pandas
df = pandas.read_csv('EnergyFirstLevel.csv')
df

df.groupby("user_screen_name").agg('count')['friends_count']

# GLI UTENTI UNIVOCI SONO SOLO 1398 SU 2039
sel_df = df[['screen_name']].drop_duplicates().shape
sel_df

# DROP_DUPLICATES LASCIA 1556 UTENTI PERCHE' ALCUNI SONO PRESENTI PIU' VOLTE CON 
# DIFFERENTI VALORI DI followers_count, statuses_count, listed_count E friends_count
friends_df = df.drop(['user_screen_name'], axis=1).drop_duplicates()
friends_df

friends_sorted_df = friends_df.sort_values(by=['screen_name','followers_count', 'statuses_count'])
friends_sorted_df

# STUDIO LA MOLTEPLICITA' ATTUALE DEGLI UTENTI
multiple_friends_df = friends_sorted_df.groupby("screen_name")\
    .agg('count')[['friends_count']]\
    .sort_values(by=['friends_count'], ascending=False)
multiple_friends_df[multiple_friends_df['friends_count']>1]

# PER OGNI UTENTE TENGO SOLO LA RIGA CON PIU' ALTO VALORE DI followers_count
idx = friends_sorted_df.groupby(['screen_name'])['followers_count'].transform(max) == friends_sorted_df['followers_count']
friends_sorted_removeduplbyFC_df = friends_sorted_df[idx]
#friends_sorted_removeduplbyFC_df

# PER OGNI UTENTE TENGO SOLO LA RIGA CON PIU' ALTO VALORE DI statuses_count
idx2 = friends_sorted_removeduplbyFC_df.groupby(['screen_name'])['statuses_count'].transform(max) == friends_sorted_removeduplbyFC_df['statuses_count']
friends_sorted_removeduplbyFCandSC_df = friends_sorted_removeduplbyFC_df[idx2]
#friends_sorted_removeduplbyFCandSC_df

# STUDIO LA MOLTEPLICITA' ATTUALE DEGLI UTENTI
multiple_friends_sorted_removeduplbyFCandSC_df = friends_sorted_removeduplbyFCandSC_df.groupby("screen_name")\
    .agg('count')[['friends_count']]\
    .sort_values(by=['friends_count'], ascending=False)
multiple_friends_sorted_removeduplbyFCandSC_df[multiple_friends_sorted_removeduplbyFCandSC_df['friends_count']>1]

friends_sorted_removeduplbyFCandSC_df[friends_sorted_removeduplbyFCandSC_df['screen_name'] == 'EU_ENV']

# PER OGNI UTENTE TENGO SOLO LA RIGA CON PIU' ALTO VALORE DI listed_count
idx3 = friends_sorted_removeduplbyFCandSC_df.groupby(['screen_name'])['listed_count'].transform(max) == friends_sorted_removeduplbyFCandSC_df['listed_count']
friends_sorted_removeduplbyFCandSCandLC_df = friends_sorted_removeduplbyFCandSC_df[idx3]
#friends_sorted_removeduplbyFCandSCandLC_df

# STUDIO LA MOLTEPLICITA' ATTUALE DEGLI UTENTI
multiple_friends_sorted_removeduplbyFCandSCandLC_df = friends_sorted_removeduplbyFCandSCandLC_df.groupby("screen_name")\
    .agg('count')[['friends_count']]\
    .sort_values(by=['friends_count'], ascending=False)
multiple_friends_sorted_removeduplbyFCandSCandLC_df[multiple_friends_sorted_removeduplbyFCandSCandLC_df['friends_count']>1]

friends_sorted_removeduplbyFCandSCandLC_df[friends_sorted_removeduplbyFCandSCandLC_df['screen_name']=='ItaSIF']

# PER OGNI UTENTE TENGO SOLO LA RIGA CON PIU' ALTO VALORE DI friends_count
idx4 = friends_sorted_removeduplbyFCandSCandLC_df.groupby(['screen_name'])['friends_count'].transform(max) == friends_sorted_removeduplbyFCandSCandLC_df['friends_count']
friends_sorted_removeduplbyFCandSCandLCandFRC_df = friends_sorted_removeduplbyFCandSCandLC_df[idx4]
#friends_sorted_removeduplbyFCandSCandLCandFRC_df

friends_sorted_removeduplbyFCandSCandLCandFRC_df.shape

unique_friends_sorted_df = friends_sorted_removeduplbyFCandSCandLCandFRC_df.sort_values(by=['followers_count','statuses_count'], ascending=False)
unique_friends_sorted_df

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>1000000].shape[0]

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>500000].shape[0]

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>250000].shape[0]

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>100000].shape[0]

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>50000].shape[0]

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>25000].shape[0]

unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>10000].shape[0]

sel_friend_by_FC = unique_friends_sorted_df[unique_friends_sorted_df['followers_count']>1000]
sel_friend_by_FC.shape[0]

sel_friends_by_SC = unique_friends_sorted_df[unique_friends_sorted_df['statuses_count']>1000]
sel_friends_by_SC.shape[0]

i1 = sel_friend_by_FC.index 
i2 = sel_friends_by_SC.index
inters_index = i1.intersection(i2)
inters_index.shape[0]

Final_friends_1st_level = unique_friends_sorted_df.loc[inters_index]
Final_friends_1st_level

Final_friends_1st_level.to_csv("friends_1st_level_filtered.csv")

Final_friends_1st_level[['screen_name','description']].to_csv("friends_1st_level_filtered_4_doccano.csv", index=False)

Final_friends_1st_level[['screen_name']].to_csv("screen_names_1st_level.txt", index=False)
